// types/penduduk.d.ts
export type Penduduk = {
  nik: string;
  nokk: string | null;
  nama: string;
  alias: string | null;
  agama: string | null;
  jenis_kelamin: string | null;
  kewarganegaraan: string;
  padukuhan: string | null;
  rt: number | null;
  rw: number | null;
  pendidikan_kk: string | null;
  pendidikan_sdt: string | null;
  pekerjaan: string | null;
  tanggal_lahir: DateTime | null;
  tempat_lahir: string | null;
  umur: number | null;
  status_kawin: string | null;
  shdk: string | null;
  gol_darah: string | null;
  nama_ayah: string | null;
  nama_ibu: string | null;
  jabatan_di_kalurahan: string;
  status_duk: string | null;
}[];

export type Individu = {
  nik: string;
  nokk: string | null;
  nama: string;
  alias: string | null;
  agama: string | null;
  jenis_kelamin: string | null;
  kewarganegaraan: string;
  padukuhan: string | null;
  rt: number | null;
  rw: number | null;
  pendidikan_kk: string | null;
  pendidikan_sdt: string | null;
  pekerjaan: string | null;
  tanggal_lahir: DateTime | null;
  tempat_lahir: string | null;
  umur: number | null;
  status_kawin: string | null;
  shdk: string | null;
  gol_darah: string | null;
  nama_ayah: string | null;
  nama_ibu: string | null;
  jabatan_di_kalurahan: string;
  status_duk: string | null;
};